start of Refactor to one main background class
